add pc checker
