+++ backend/infrahub/proposed_change/tasks.py
@@ -102,7 +102,7 @@ from infrahub.workflows.catalogue import (
 from infrahub.workflows.utils import add_tags
 
 from .branch_diff import get_diff_summary_cache, get_modified_kinds
-from .revoke_approvals import do_check_pc_has_enough_approvals
+from .checker import can_merge_proposed_change
 
 if TYPE_CHECKING:
     from infrahub_sdk.client import InfrahubClient
@@ -171,15 +171,8 @@ async def merge_proposed_change(
     async with database.start_session() as db:
         log.info("Validating if all conditions are met to merge the proposed change")
 
-        nb_approvals_required = config.SETTINGS.policy.required_proposed_change_approvals
-        if nb_approvals_required > 0:
-            error_message = await do_check_pc_has_enough_approvals(
-                nb_approvals_required=nb_approvals_required,
-                proposed_change=proposed_change,  # type: ignore[arg-type]
-                db=db,
-            )
-            if error_message:
-                return Failed(message=error_message)
+        if error_message := await can_merge_proposed_change(proposed_change=proposed_change, db=db):
+            return Failed(message=error_message)
 
         source_branch = await Branch.get_by_name(db=db, name=proposed_change.source_branch.value)
         validations = await proposed_change.validations.get_peers(db=db, peer_type=CoreValidator)
